Move the Keyboard layout menu item to the configure menu
