+++ src/esp/components/ptrs/ptrs.cpp
@@ -8,17 +8,15 @@
 
 #define MENU_CONFIGURE 0
 #define MENU_ROMS 1
-#define MENU_KBLAYOUT 2
-#define MENU_STATUS 3
-#define MENU_RESET 4
-#define MENU_HELP 5
-#define MENU_EXIT 6
+#define MENU_STATUS 2
+#define MENU_RESET 3
+#define MENU_HELP 4
+#define MENU_EXIT 5
 
 
 static menu_item_t main_menu_items[] = {
   MENU_ITEM(MENU_CONFIGURE, "Configure"),
   MENU_ITEM(MENU_ROMS, "ROMs"),
-  MENU_ITEM(MENU_KBLAYOUT, "Keybord Layout"),
   MENU_ITEM(MENU_STATUS, "Status"),
   MENU_ITEM(MENU_RESET, "Reset Settings"),
   MENU_ITEM(MENU_HELP, "Help"),
@@ -54,9 +52,6 @@ void configure_pocket_trs(bool is_80_cols)
       break;
     case MENU_STATUS:
       ptrs_status();
-      break;
-    case MENU_KBLAYOUT:
-      configure_Keyboard_Layout();
       break;
     }
     show_from_left = true;
